java/bluemix-todo-app/src/test/java/net/bluemix/todo/model/CloudantReducedRowTest.java
/*
 * Copyright IBM Corp. 2014
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package net.bluemix.todo.model;

import static org.junit.Assert.assertEquals;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;
public class CloudantReducedRowTest {

  @Before
  public void setUp() throws Exception {}

  @After
  public void tearDown() throws Exception {}

  @Test
  public void testGetKey() {
    CloudantReducedRow row = new CloudantReducedRow();
    row.setKey("123");
    assertEquals("123", row.getKey());
  }

  @Test
  public void testGetValue() {
    CloudantReducedRow row = new CloudantReducedRow();
    row.setValue(123);
    assertEquals(123, row.getValue());
  }
}

